+++ OSDEPYM/src/main/java/com/osdepym/hibernate/dao/BeneficiarioDAO.java
@@ -6,5 +6,5 @@ import com.osdepym.exception.CustomException;
 import com.osdepym.hibernate.entity.Beneficiario;
 
 public interface BeneficiarioDAO extends Dao<Beneficiario>{
-	public List<Beneficiario> getBeneficiariosByIdAfiliado(int idAfiliado) throws CustomException;
+	public List<Beneficiario> getBeneficiariosByIdAfiliado(Long idAfiliado) throws CustomException;
 }
